mv bubble tea logic to module, mv main to gotp
